Move out data traits into their own file
